% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{associate_phenotype_modules}
\alias{associate_phenotype_modules}
\title{Modules phenotpic association}
\usage{
associate_phenotype_modules(eigengenes, phenotypes)
}
\arguments{
\item{eigengenes}{Matrix of igengenes provided by the output of modules_detection.}

\item{phenotypes}{Matrix of phenotypes for each sample to associate.
TODO finish}
}
\value{
TODO
}
\description{
Compute the correlation between all modules and the phenotypic variables
}
\details{
TODO
}
\examples{
#TODO
}
