% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{bio_enrich}
\alias{bio_enrich}
\title{Modules enrichment}
\usage{
bio_enrich(module, custom_gmt = NULL, ...)
}
\arguments{
\item{TODO}{finish}
}
\value{
TODO
}
\description{
Enrich genes list from modules.
}
\details{
TODO
}
\examples{
#TODO
}
