% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{build_graph_from_sq_mat}
\alias{build_graph_from_sq_mat}
\title{Return graph from squared matrix network}
\usage{
build_graph_from_sq_mat(sq_mat)
}
\arguments{
\item{sq_mat}{matrix or data.frame, squared matrix representing}
}
\description{
Takes a squared matrix containing the pairwise similarity scores for each gene and return a igraph object.
}
