% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{check_gost}
\alias{check_gost}
\title{Check if it's a gost result}
\usage{
check_gost(gost_result)
}
\arguments{
\item{gost_result}{list of gprofiler2::gost result}
}
\description{
Take a list that should be a gost result and check if format is good.
}
