% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{compare_modules}
\alias{compare_modules}
\title{Compare modules topology between conditions}
\usage{
compare_modules(data_expr_list, net_list, cor_list = NULL, modules_list,
  ref = names(data_expr_list)[1], cor_func = c("pearson", "spearman",
  "bicor", "other"), your_func = NULL, n_perm = 10000,
  comparison_modules = c("unpreserved", "preserved", "one or the other"),
  pvalue_th = 0.05, n_threads = NULL, ...)
}
\arguments{
\item{data_expr_list}{list of data_expr, list of expression data with genes as column and samples as row.}

\item{net_list}{list of networks, list of square tables representing connectivity between
each genes as returned by build_net.}

\item{cor_list}{list of matrices and/or data.frames, list of square tables representing correlation between
each gene. Must be the same used to create networks in \code{\link{net_list}}. If NULL, will be re-calculated
according to \code{cor_func}.}

\item{modules_list}{list of modules or nested list of modules, list of modules in one condition (will be considered as the one from reference) or a condition
named list with list of modules built in each one.}

\item{ref}{string, condition name to be used as reference for permutation tests or "cross comparison" if you want to compare
each condition with the other as reference. Default will be the name of the first element
in data_expr_list.}

\item{cor_func}{string, name of the correlation function to be used. Must be one of "pearson", "spearman",
"bicor", "other". If "other", your_func must be provided}

\item{your_func}{function returning correlation values. Final values must be in [-1;1]}
}
\description{
Take modules built from multiples conditions and search for preservation or non-preservation of them against one or mutliple
conditions of reference. Use 7 topological features to perform the differents test, and use permutation to validate results.
}
\details{
\describe{
  \item{To avoid recalculation, correlations matrices can be obtain by setting \code{keep_cor_mat} in \code{\link{net_list}} to TRUE.}
  \item{Description of the 7 topological features used for preservation testing is available in \code{\link[NetRep]{modulePreservation}}.}
}
}
