% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{detect_modules}
\alias{detect_modules}
\title{Modules detection in a network}
\usage{
detect_modules(data_expr, net, min_module_size = min(20,
  ncol(data_expr)/2), merge_close_modules = TRUE,
  merge_cut_height = 0.25, detailled_result = FALSE, ...)
}
\arguments{
\item{data_expr}{matrix of expression data with genes as column and samples as row.}

\item{tom}{TODO finish}
}
\value{
TODO
}
\description{
Detect the modules by hierarchical clustering .
}
\details{
TODO
}
\examples{
#TODO
}
