% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_data_expr}
\alias{.check_data_expr}
\title{Run checks on an object to test if it's a data_expr}
\usage{
.check_data_expr(data_expr)
}
\arguments{
\item{data_expr}{matrix or data.frame, expression data with genes as
column and samples as row.}
}
\value{
Throw an error if doesn't correspond
}
\description{
Check an object to be a data.frame or a matrix compatible of genes and
samples.
}
