% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_gost}
\alias{.check_gost}
\title{Run checks on an object to test if it's a gost result}
\usage{
.check_gost(gost_result)
}
\arguments{
\item{gost_result}{list, gprofiler2::gost result}
}
\value{
Throw an error if doesn't correspond
}
\description{
Take a list that should be a gost result and check if format is good.
}
