% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_network}
\alias{.check_network}
\title{Run checks on an object to test if it's a network}
\usage{
.check_network(network)
}
\arguments{
\item{network}{matrix or data.frame, object to test to be a network}
}
\value{
Throw an error if doesn't correspond
}
\description{
Check content of a given object to determine if it's a network, meaning a
squared matrix of similarity
score between genes.
}
