% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrep_functions.R
\name{.contingencyTable}
\alias{.contingencyTable}
\title{Calculate a contigency table of module overlap between datasets}
\usage{
.contingencyTable(modAssignments, mods, tiNodelist)
}
\arguments{
\item{modAssignments}{a list where the first element is the
'moduleAssignments' vector in the discovery dataset, and the second element
is the 'moduleAssignments vector in the test dataset.}

\item{mods}{the 'modules' vector for the discovery dataset.}

\item{tiNodelist}{a vector of node IDs in the test dataset.}
}
\value{
A list containing a contigency table, a vector of the proportion of
 nodes present in the test dataset for each module, a vector containing
 the number of nodes present in the test dataset for each module, a vector
 of the node names present in both the discovery and test datasets, a vector
 of modules that are both requested and have nodes present in the test
 dataset, and the \code{modAssignments} vector containing only nodes present
 in the test dataset.
}
\description{
Calculate a contigency table of module overlap between datasets
}
