% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{.cor_func_match}
\alias{.cor_func_match}
\title{Match a correlation function based on a name}
\usage{
.cor_func_match(cor_func = c("pearson", "spearman", "bicor"))
}
\arguments{
\item{cor_func}{string of the name of the correlation to be use}
}
\value{
A function corresponding to the correlation required
}
\description{
Translate a function name into an R function.
}
