% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{filter_low_var}
\alias{filter_low_var}
\title{Filtering on low variating genes}
\usage{
filter_low_var(data_expr, pct = 0.8, type = c("mean", "median", "mad"))
}
\arguments{
\item{data_expr}{matrix of data to be filtered, with genes as column and samples as row}

\item{pct}{number in 0 and 1 specifying the percentage of gene to keep}

\item{type}{string which should be either "mean" or "median"}
}
\description{
Remove low variating genes based on the percentage given and the type of variation specified.
}
