% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{filter_RNA_seq}
\alias{filter_RNA_seq}
\title{Filtering of low counts}
\usage{
filter_RNA_seq(data_expr, min_count = 5)
}
\arguments{
\item{min_count}{}
}
\description{
Remove genes with low counts in matrix coming from RNA-seq data.
}
\details{
Low counts in RNA-seq can bring noise to gene co-expression module building, so filtering them help to improve quality.
}
