% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_graph}
\alias{get_graph}
\title{Getting a graph from network matrix}
\usage{
get_graph(tom)
}
\arguments{
\item{modules_phenotype}{List form modules_phenotype function output
TODO finish}
}
\value{
TODO
}
\description{
Transform a network matrix into a graph vertices and edges
}
\details{
TODO
}
\examples{
#TODO


}
