% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_palette}
\alias{gg_palette}
\title{Mimicking ggplot palette
Source : https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
\usage{
gg_palette(n)
}
\arguments{
\item{n}{integer, number of colors wanted}
}
\value{
character vector, haxadecimal colors of length n
}
\description{
Mimicking ggplot palette
Source : https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
