% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{modules_enrichment}
\alias{modules_enrichment}
\title{Modules enrichment}
\usage{
modules_enrichment(modules, custom_gmt = NULL, ...)
}
\arguments{
\item{data_expr}{matrix of expression data with genes as column and samples as row.}

\item{tom}{TODO finish}
}
\value{
TODO
}
\description{
Enrich genes list from modules.
}
\details{
TODO
}
\examples{
TODO
}
