% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservation.R
\name{modules_preservation}
\alias{modules_preservation}
\title{Checking preservation of modules between one or more conditions}
\usage{
modules_preservation(data_expr_list, net_list, modules_list, reference,
  test)
}
\arguments{
\item{conditions}{List form modules_phenotype function output
TODO finish}
}
\value{
TODO
}
\description{
Detect which modules are preserved from a condition to another
}
\details{
TODO
}
\examples{
TODO

}
