% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrep_functions.R
\name{orderAsNumeric}
\alias{orderAsNumeric}
\title{Order the module vector numerically}
\usage{
orderAsNumeric(vec)
}
\arguments{
\item{vec}{module vector to order}
}
\value{
the order of the vector
}
\description{
The module assingments may be numeric, but coded as characters.
}
\keyword{internal}
