% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{plot_comparison_stats}
\alias{plot_comparison_stats}
\title{Heatmap of comparison statistics}
\usage{
plot_comparison_stats(comparison_pvalues, pvalue_th = 0.05,
  low_color = "#031643", mid_color = "#A0A3D3",
  high_color = "#FFFFFF")
}
\arguments{
\item{comparison_pvalues}{matrix or data.frame, table containing the p values for the statistics on each module}

\item{pvalue_th}{decimal, threshold of pvalue below which statistics are considered as significant}

\item{low_color, mid_color, high_color}{string, color to use as lower, middle, and higher end of the legend. Can
either be the color name or hexadecimal code (e.g.: “red” or “#FF1234” )}
}
\description{
Plot heatmap of p values for the module comparison statistics evaluated through the permutation test.
}
