% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot module from bio_enrich}
\usage{
plot_enrichment(enrich_output, modules = "all", sources = "all",
  colorblind = TRUE, custom_palette = NULL, ...)
}
\arguments{
\item{enrich_output}{list, bio_enrich result which are in fact gprofiler2::gost output.}

\item{modules}{string or vector of characters designing the modules to plot. "all" by default to plot every module.}

\item{sources}{string or vector of characters designing the sources to plot. "all" by default to plot every source.}

\item{...}{any other parameter you can provide to gprofiler2::gostplot}
}
\description{
Wrapper of the gprofiler2::gostplot function. Adding support of colorblind palet and selection of subsets if
initial multiple query, and/or sources to plot.
}
