% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{plot_modules_merge}
\alias{plot_modules_merge}
\title{Merging modules' plot}
\usage{
plot_modules_merge(modules)
}
\arguments{
\item{TODO}{finish}
}
\value{
TODO
}
\description{
Plot a bipartite graph to see in which modules all modules have been merged
}
\details{
TODO
}
\examples{
#TODO
}
