% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{plot_modules_phenotype}
\alias{plot_modules_phenotype}
\title{Heatmap of modules phenotpic association}
\usage{
plot_modules_phenotype(modules_phenotype, signif_th = 0.05)
}
\arguments{
\item{modules_phenotype}{List form modules_phenotype function output
TODO finish}
}
\value{
TODO
}
\description{
Plot a heatmap of the correlation between all modules and the phenotypic variables and the p value associated
}
\details{
TODO
}
\examples{
#TODO
}
