% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Muting a function}
\usage{
quiet(func)
}
\arguments{
\item{func}{Function who need to be muted.}
}
\value{
Nothing, just mute the called function
}
\description{
Prevent a function to output multiple message.
Source
}
