% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_integration.R
\name{associate_phenotype}
\alias{associate_phenotype}
\title{Modules phenotpic association}
\usage{
associate_phenotype(eigengenes, phenotypes)
}
\arguments{
\item{eigengenes}{matrix or data.frame, eigengenes of the modules. Provided by the output of modules_detection.}

\item{phenotypes}{matrix or data.frame, phenotypes for each sample to associate.}
}
\description{
Compute the correlation between all modules and the phenotypic variables
}
