% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{.check_data_expr}
\alias{.check_data_expr}
\title{Run all checks on expression data}
\usage{
.check_data_expr(data_expr)
}
\arguments{
\item{data_expr}{matrix or data.frame, expression data with genes as column and samples as row.}
}
\value{
A boolean, indicate if all checks were good
}
\description{
Check that a data_expr is a matrix or data frame conform for process by GWENA
}
