% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{.check_data_expr}
\alias{.check_data_expr}
\title{Run all checks on expression data}
\usage{
.check_data_expr(data_expr)
}
\value{
A boolean, indicate if all checks were good
}
\description{
Check that it is a matrix or data frame,
}
