% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{filter_RNA_seq}
\alias{filter_RNA_seq}
\title{Filtering of low counts}
\usage{
filter_RNA_seq(data_expr, min_count = 5, method = c("at least one",
  "mean", "all"))
}
\arguments{
\item{data_expr}{matrix or data.frame, table of expression values (either microarray or RNA-seq),
with genes as column and samples as row.}

\item{min_count}{integer, minimal number of count to be considered in method.}

\item{method}{string, name of the method for filtering. Must be one of "at least one", "mean", or " all"}
}
\description{
Keeping genes with at least one sample with count above min_count in RNA-seq data.
}
\details{
Low counts in RNA-seq can bring noise to gene co-expression module building, so filtering them
help to improve quality.
}
