% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_building.R
\name{get_fit.cor}
\alias{get_fit.cor}
\title{Calculating best fit of a power low on correlation matrix computed on expression data}
\usage{
get_fit.cor(cor_mat, fit_cut_off = 0.9, network_type = c("unsigned",
  "signed", "signed hybrid", "none"), ...)
}
\arguments{
\item{cor_mat}{matrix or data.frame of genes correlation.}

\item{network_type}{TODO finish params}

\item{power_cut_off}{integer by which R^2 (coefficient of determination) will be thresholded.}
}
\value{
A list containing power of the law for best fit, fit table, and metadata about the arguments used.
}
\description{
Adjust a correlation matrix depending of the type of network, then try to parameter a power law for best fit
}
\examples{
TODO Add example

}
\author{
Gwenaëlle Lemoine <lemoine.gwenaelle@gmail.com>
}
