% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{get_fit.expr}
\alias{get_fit.expr}
\title{Calculating best fit of a power low on expression data}
\usage{
get_fit.expr(data_expr, fit_cut_off = 0.9, cor_func = c("pearson",
  "spearman", "bicor", "other"), your_func = NULL,
  network_type = c("unsigned", "signed", "signed hybrid", "none"), ...)
}
\arguments{
\item{data_expr}{matrix of data only normalized for constructor specificities, with genes as column and samples as row.}

\item{cor_func}{TODO}

\item{your_func}{TODO}

\item{network_type}{TODO}

\item{...}{TODO filling info on params}

\item{power_cut_off}{integer by which R^2 (coefficient of determination) will be thresholded}
}
\value{
A list containing power of the law for best fit, fit table, and metadata about the arguments used.
}
\description{
Computes correlation matrix of the gene expression data, adjust it depending of the type of network, then try to
parameter a power law for best fit
}
\examples{
#TODO write example

}
