% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_expr}
\alias{gtex_expr}
\title{Transcriptomic muscle data from GTEx consorsium RNA-seq data}
\format{A data frame with 50 rows (samples) and 15000 columns (genes)}
\source{
\url{https://gtexportal.org/home/datasets}
}
\usage{
gtex_expr
}
\description{
A subset of GTEx RNA-seq dataset containing read counts collapsed to gene level.
}
\keyword{datasets}
