% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_gost}
\alias{is_gost}
\title{Determine if an object is a gost object}
\usage{
is_gost(gost_result)
}
\arguments{
\item{gost_result}{list, gprofiler2::gost result}
}
\value{
list, a boolean as first element and in second element NULL or the reason why boolean is
set to FALSE
}
\description{
Check content of a given object to determine if it's a gost object
}
