% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_building.R
\name{net_building}
\alias{net_building}
\title{Network build and modules detection}
\usage{
net_building(data_expr, cor_func = c("pearson", "spearman", "bicor",
  "other"), your_func = NULL, power_value = NULL, fit_cut_off = 0.9,
  network_type = c("unsigned", "signed", "signed hybrid", "none"),
  tom_type = c("unsigned", "signed", "signed Nowick"),
  min_module_size = min(20, ncol(data_expr)/2),
  merge_cut_height = 0.25, save_adjacency = FALSE, save_tom = FALSE,
  n_threads = 0, detailled_result = FALSE, ...)
}
\arguments{
\item{data_expr}{matrix of expression data with genes as column and samples as row.}

\item{cor_func}{string specifying correlation function to be used. Must be one of pearson, spearman, bicor.
TODO finish}

\item{power_value}{soft-thresolding, meaning power which will be applied to the adjacency matrix.}

\item{power_cut_off}{if no power provided, computes power and thresold it by this value.}
}
\value{
list containing
}
\description{
Compute the adjacency matrix, then the TOM to build the network. Than detect the modules by hierarchical clustering and thresholding
}
\details{
\item{which should have been adequatly normalized and filtered if pertinant. ATTENTION : it is not recommended to filter by differential
 expression (cf. \link(https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/faq.html)}
}
\examples{

}
