% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_and_modules.R
\name{plot_expression_profiles}
\alias{plot_expression_profiles}
\title{Plot all modules expression profiles}
\usage{
plot_expression_profiles(data_expr, modules)
}
\arguments{
\item{TODO}{finish}
}
\value{
TODO
}
\description{
f
}
\details{
TODO
}
\examples{
#TODO
}
