% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{plot_module}
\alias{plot_module}
\title{Plot co-expression network}
\usage{
plot_module(graph_module, hubs = NULL, weight_th = 0.2,
  enrichment = NULL, title = "Module", degree_node_scaling = TRUE,
  node_scaling = 1, edge_scaling = 5, layout = "auto", ...)
}
\arguments{
\item{graph_module}{igraph object, module to plot.}

\item{hubs}{character vector or numeric vector with names, optionnal, vector of gene names or vector of numeric
values named with gene names}

\item{weight_th}{decimal, weight threshold under or equal to which edges will be removed}

\item{enrichment}{list, representing a gost object}

\item{title}{string, main title that will be displayed on the plot}

\item{degree_node_scaling}{boolean, indicate if node size should represent the degree of this node}

\item{node_scaling}{numeric, scaling factor by whihch the node size will be scalled}

\item{edge_scaling}{numeric, scaling factor by whihch the edge width will be scalled}

\item{layout}{numeric matrix or function or string, numeric matrix for nodes coordinates, or function for layout,
or name of a layout function available in \code{igraph}. Default "auto" will choose the best layout depending on
the graph. For more information, see \code{\link{[igraph]{igraph.plotting}}}}

\item{...}{any other parameter compatible with the \code{\link{[igraph]{plot.igraph}}} function}
}
\description{
Display a graph representing the co-expression network and different informations like hubs, enrichments
}
