% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{plot_module}
\alias{plot_module}
\title{Plot co-expression network}
\usage{
plot_module(graph_module, hubs = NULL, weight_th = 0.2,
  enrichment = NULL, title = "Module", degree_node_scaling = TRUE,
  edge_scaling = 5, node_scaling = 1, ...)
}
\arguments{
\item{hubs}{vector, optionnal,}
}
\description{
Display a graph representing the co-expression network and different informations like hubs, enrichments
}
